chore: Moved refresh endpoint to auth controller
